first commit, assignPlayerNumber implemented
