+++ client/src/App.jsx
@@ -1,10 +1,13 @@
 import { EthProvider } from "./contexts/EthContext";
 import { GameBoard } from "./components/Board/GameBoard";
+import PlayerNumber from "./components/PlayerNumber/PlayerNumber";
+
 function App() {
   return (
     <EthProvider>
       <div id="App">
         <div className="container">
+        <PlayerNumber />
         <GameBoard />
         </div>
       </div>
