--- client/src/components/Demo/Cta.jsx
@@ -1,12 +1,0 @@
-function Cta() {
-  return (
-    <p>
-      Try changing&nbsp;
-      <span className="code">value</span>
-      &nbsp;in&nbsp;
-      <span className="code">SimpleStorage</span>.
-    </p>
-  );
-}
-
-export default Cta;
