--- client/src/components/Demo/Desc.jsx
@@ -1,20 +1,0 @@
-function Desc() {
-  return (
-    <>
-      <p>
-        Take a look at <span className="code">client/src/contexts/EthContext</span>.
-        This context maintains a global state and provides web3.js functionalities
-        to the rest of the app.
-      </p>
-      <p>
-        Feel free to remove any component or styling that you don't need, and
-        extend <span className="code">EthContext</span> to your dapp's needs.
-      </p>
-      <p>
-        Happy hacking!
-      </p>
-    </>
-  );
-}
-
-export default Desc;
