--- client/src/components/Demo/NoticeNoArtifact.jsx
@@ -1,10 +1,0 @@
-function NoticeNoArtifact() {
-  return (
-    <p>
-      ⚠️ Cannot find <span className="code">SimpleStorage</span> contract artifact.
-      Please complete the above preparation first, then restart the react dev server.
-    </p>
-  );
-}
-
-export default NoticeNoArtifact;
