--- client/src/components/Demo/NoticeWrongNetwork.jsx
@@ -1,9 +1,0 @@
-function NoticeWrongNetwork() {
-  return (
-    <p>
-      ⚠️ MetaMask is not connected to the same network as the one you deployed to.
-    </p>
-  );
-}
-
-export default NoticeWrongNetwork;
