--- client/src/components/Demo/Title.jsx
@@ -1,5 +1,0 @@
-function Title() {
-  return <h2>See it in action</h2>;
-}
-
-export default Title;
