+++ client/src/contexts/EthContext/EthProvider.jsx
@@ -30,7 +30,7 @@ function EthProvider({ children }) {
   useEffect(() => {
     const tryInit = async () => {
       try {
-        const artifact = require("../../contracts/SimpleStorage.json");
+        const artifact = require("../../contracts/Monopoly.json");
         init(artifact);
       } catch (err) {
         console.error(err);
